kortex/requirements.txt
aiohttp
anthropic
flask[async]
flask-cors
google-generativeai
openai
PyPDF2
python-dotenv
